feat[#131]: KTX & Report changed exceptions
