+++ src/main/java/com/itaxi/server/report/application/ReportService.java
@@ -33,40 +33,30 @@ public class ReportService {
 
     @Transactional
     public ReportResponse createReport(AddReportDto dto) {
-        // 신고한 사람과 당한 사람 둘다 존재하는지 체크
         Optional<Member> writer = memberRepository.findMemberByUid(dto.getWriterUid());
         Optional<Member> reportedMember = memberRepository.findMemberByUid(dto.getReportedUid());
         Report report = null;
         if (writer.isPresent() && reportedMember.isPresent()) {
             AddReportMemberDto reportMemberDto = new AddReportMemberDto(dto, writer.get(), reportedMember.get());
-            // 신고 당한 사람의 페널티를 +1
             Member member = reportedMember.get();
             member.setPenalty(member.getPenalty() + 1);
-            // 멤버와 리포트를 저장
             memberRepository.save(member);
             report = reportRepository.save(reportMemberDto.toEntity());
         } else {
             throw new MemberNotFoundException(HttpStatus.INTERNAL_SERVER_ERROR);
         }
 
-        ReportResponse response = new ReportResponse(
-                report.getId(),
-                new MemberResponse(report.getWriter().getId(), report.getWriter().getName()),
-                new MemberResponse(report.getReportedMember().getId(), report.getReportedMember().getName()),
-                        report.getDate(), report.getTitle(), report.getContent());
+        ReportResponse response = new ReportResponse(report);
 
         return response;
     }
 
     @Transactional
     public List<ReportGetResDto> getReport(String uid) {
-        // 해당 uid를 가진 멤버 존재 확인
         Optional<Member> member = memberRepository.findMemberByUid(uid);
         if(!member.isPresent())
             throw new MemberNotFoundException(HttpStatus.INTERNAL_SERVER_ERROR);
-        // 멤버 아이디 가져오기
         Long id = member.get().getId();
-        // 자신의 신고내역을 보내준다
         List<Report> posts = reportRepository.findAllByWriterId(id);
         List<ReportGetResDto> results = posts.stream().map(m -> new ReportGetResDto(m)).collect(Collectors.toList());
 
@@ -77,23 +67,19 @@ public class ReportService {
     public String updateReport(Long reportId, UpdateReportDto dto) {
         Report reportInfo = null;
         Member memberInfo = null;
-        // report 존재 체크
         Optional<Report> report = reportRepository.findById(reportId);
         if (report.isPresent()) {
             reportInfo = report.get();
         } else {
             throw new ReportNotFoundException(HttpStatus.BAD_REQUEST);
         }
-        // member 가져와서 존재 체크
         Optional<Member> member = memberRepository.findMemberByUid(dto.getUid());
         if (member.isPresent()) {
             memberInfo = member.get();
         } else {
             throw new MemberNotFoundException(HttpStatus.BAD_REQUEST);
         }
-        // 멤버가 작성한 사람인지 체크
         if (reportInfo.getWriter().getId() == memberInfo.getId()) {
-            // 작성한 사람이면 content랑 title 변경하고 저장
             reportInfo.setContent(dto.getContent());
             reportInfo.setTitle(dto.getTitle());
             reportRepository.save(reportInfo);
@@ -108,23 +94,19 @@ public class ReportService {
     public String deleteReport(Long reportId, String uid) {
         Report reportInfo = null;
         Member memberInfo = null;
-        // report 존재 체크
         Optional<Report> report = reportRepository.findById(reportId);
         if (report.isPresent()) {
             reportInfo = report.get();
         } else {
             throw new ReportNotFoundException(HttpStatus.BAD_REQUEST);
         }
-        // member 가져와서 존재 체크
         Optional<Member> member = memberRepository.findMemberByUid(uid);
         if (member.isPresent()) {
             memberInfo = member.get();
         } else {
             throw new MemberNotFoundException(HttpStatus.BAD_REQUEST);
         }
-        // 멤버가 작성한 사람인지 체크
         if (reportInfo.getWriter().getId() == memberInfo.getId()) {
-            // 작성한 사람이면 content랑 title 변경하고 저장
             reportInfo.setDeleted(true);
             reportRepository.save(reportInfo);
         } else {
